+++ src/gateways/InclusiveGateway.js
@@ -1,6 +1,6 @@
 import Activity from '../activity/Activity';
 import {ActivityError} from '../error/Errors';
-import {cloneContent} from '../messageHelper';
+import {cloneContent, cloneMessage} from '../messageHelper';
 
 export default function InclusiveGateway(activityDef, context) {
   return Activity(InclusiveGatewayBehaviour, activityDef, context);
@@ -21,47 +21,24 @@ export function InclusiveGatewayBehaviour(activity) {
     const content = cloneContent(executeMessage.content);
     if (!outboundSequenceFlows.length) return complete();
 
-    let conditionMet, defaultFlow, evaluateError;
-    const outbound = content.outbound = [];
+    return activity.evaluateOutbound(cloneMessage(executeMessage), false, complete);
 
-    for (let i = 0; i < outboundSequenceFlows.length; i++) {
-      const flow = outboundSequenceFlows[i];
+    function complete(err, outbound) {
+      if (err) return broker.publish('execution', 'execute.error', cloneContent(content, {error: err}));
+      if (!outbound) return broker.publish('execution', 'execute.completed', content);
 
-      if (flow.isDefault) {
-        defaultFlow = flow;
-        continue;
+      const taken = outbound.find(({action}) => action === 'take');
+      if (!taken) {
+        const error = new ActivityError(`<${id}> no conditional flow taken`, executeMessage);
+        logger.error(`<${id}>`, err);
+        return broker.publish('execution', 'execute.error', {...content, error});
       }
 
-      if (flow.evaluateCondition(executeMessage, onEvaluateError)) {
-        conditionMet = true;
-        outbound.push({id: flow.id, action: 'take'});
-      } else {
-        if (evaluateError) return broker.publish('execution', 'execute.error', cloneContent(content, {error: evaluateError}));
-        outbound.push({id: flow.id, action: 'discard'});
+      if (taken.isDefault) {
+        logger.debug(`<${id}> take default flow <${taken.id}>`);
       }
-    }
 
-    if (defaultFlow) {
-      if (conditionMet) {
-        outbound.push({id: defaultFlow.id, action: 'discard'});
-      } else {
-        logger.debug(`<${id}> take default flow`);
-        outbound.push({id: defaultFlow.id, action: 'take'});
-      }
-    } else if (!conditionMet) {
-      const err = new ActivityError(`<${id}> no conditional flow taken`, executeMessage);
-      logger.error(`<${id}>`, err);
-      return broker.publish('execution', 'execute.error', cloneContent(content, {error: err}));
-    }
-
-    return complete();
-
-    function complete() {
-      broker.publish('execution', 'execute.completed', cloneContent(content));
-    }
-
-    function onEvaluateError(err) {
-      evaluateError = err;
+      return broker.publish('execution', 'execute.completed', {...content, outbound});
     }
   }
 }