+++ test/feature/gateway-feature.js
@@ -27,7 +27,7 @@ Feature('Gateway', () => {
           <inclusiveGateway id="decisions" default="defaultFlow" />
           <sequenceFlow id="defaultFlow" sourceRef="decisions" targetRef="end1" />
           <sequenceFlow id="condFlow1" sourceRef="decisions" targetRef="end2">
-            <conditionExpression xsi:type="tFormalExpression" language="javascript">this.environment.variables.condition.var1</conditionExpression>
+            <conditionExpression xsi:type="tFormalExpression" language="javascript">next(null, this.environment.variables.condition.var1)</conditionExpression>
           </sequenceFlow>
           <sequenceFlow id="condFlow2" sourceRef="decisions" targetRef="end3">
             <conditionExpression xsi:type="tFormalExpression">\${environment.variables.condition2}</conditionExpression>
@@ -123,7 +123,7 @@ Feature('Gateway', () => {
           <exclusiveGateway id="decisions" default="defaultFlow" />
           <sequenceFlow id="defaultFlow" sourceRef="decisions" targetRef="end1" />
           <sequenceFlow id="condFlow1" sourceRef="decisions" targetRef="end2">
-            <conditionExpression xsi:type="tFormalExpression" language="javascript">this.environment.variables.condition.var1</conditionExpression>
+            <conditionExpression xsi:type="tFormalExpression" language="javascript">next(null, this.environment.variables.condition.var1)</conditionExpression>
           </sequenceFlow>
           <sequenceFlow id="condFlow2" sourceRef="decisions" targetRef="end3">
             <conditionExpression xsi:type="tFormalExpression">\${environment.variables.condition2}</conditionExpression>
@@ -370,6 +370,111 @@ Feature('Gateway', () => {
 
     Then('run completes', () => {
       return end;
+    });
+  });
+
+  Scenario('A process with an exclusive gateway with asynchronous conditions', () => {
+    let definition;
+    Given('one default flow, second flow with async script condition, and a third with expression', async () => {
+      const source = `
+      <?xml version="1.0" encoding="UTF-8"?>
+      <definitions id="Definitions_1" xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+        <process id="mainProcess" isExecutable="true">
+          <startEvent id="start" />
+          <sequenceFlow id="flow1" sourceRef="start" targetRef="decisions" />
+          <exclusiveGateway id="decisions" default="defaultFlow" />
+          <sequenceFlow id="defaultFlow" sourceRef="decisions" targetRef="end1" />
+          <sequenceFlow id="condFlow1" sourceRef="decisions" targetRef="end2">
+            <conditionExpression xsi:type="tFormalExpression" language="javascript">this.environment.services.evaluateRule('my-rule', this.environment.variables, next);</conditionExpression>
+          </sequenceFlow>
+          <sequenceFlow id="condFlow2" sourceRef="decisions" targetRef="end3">
+            <conditionExpression xsi:type="tFormalExpression">\${environment.variables.condition2}</conditionExpression>
+          </sequenceFlow>
+          <endEvent id="end1" />
+          <endEvent id="end2" />
+          <endEvent id="end3" />
+        </process>
+      </definitions>`;
+
+      const context = await testHelpers.context(source);
+      definition = Definition(context);
+    });
+
+    When('definition is ran with falsy second and first condition script', () => {
+      definition.environment.variables.condition = {var1: false};
+      definition.environment.services.evaluateRule = async function evaluateRule(name, variables, callback) {
+        return new Promise((resolve) => {
+          resolve(variables.condition.var1);
+        }).then((result) => {
+          callback(null, result);
+        }).catch((err) => {
+          callback(err);
+        });
+      };
+      definition.run();
+    });
+
+    Then('default flow is taken', () => {
+      expect(definition.getActivityById('end1').counters).to.have.property('taken', 1);
+    });
+
+    And('the other two discarded', () => {
+      expect(definition.getActivityById('end2').counters).to.have.property('discarded', 1);
+      expect(definition.getActivityById('end3').counters).to.have.property('discarded', 1);
+    });
+
+    When('definition is ran with truthy second condition script', () => {
+      definition.environment.variables.condition = {var1: true};
+      definition.run();
+    });
+
+    Then('default flow is discarded', () => {
+      expect(definition.getActivityById('end1').counters).to.have.property('taken', 1);
+      expect(definition.getActivityById('end1').counters).to.have.property('discarded', 1);
+    });
+
+    And('the second flow is taken', () => {
+      expect(definition.getActivityById('end2').counters).to.have.property('taken', 1);
+      expect(definition.getActivityById('end2').counters).to.have.property('discarded', 1);
+    });
+
+    And('the third flow is discarded', () => {
+      expect(definition.getActivityById('end3').counters).to.have.property('discarded', 2);
+    });
+
+    When('definition is ran with truthy second and third condition script', () => {
+      definition.environment.variables.condition = {var1: true};
+      definition.environment.variables.condition2 = true;
+      definition.run();
+    });
+
+    Then('default flow is discarded', () => {
+      expect(definition.getActivityById('end1').counters).to.have.property('taken', 1);
+      expect(definition.getActivityById('end1').counters).to.have.property('discarded', 2);
+    });
+
+    And('the second flow is taken', () => {
+      expect(definition.getActivityById('end2').counters).to.have.property('taken', 2);
+      expect(definition.getActivityById('end2').counters).to.have.property('discarded', 1);
+    });
+
+    And('the third flow is still discarded', () => {
+      expect(definition.getActivityById('end3').counters).to.have.property('taken', 0);
+      expect(definition.getActivityById('end3').counters).to.have.property('discarded', 3);
+    });
+
+    let error;
+    When('definition is ran with second flow type error', () => {
+      definition.environment.variables.condition = undefined;
+      definition.environment.variables.condition2 = true;
+      definition.once('error', (err) => {
+        error = err;
+      });
+      definition.run();
+    });
+
+    Then('run is errored', () => {
+      expect(error).to.be.ok;
     });
   });
 });